pyproject.toml
[project]
name = "gesture-driven-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mediapipe>=0.10.35",
    "opencv-python>=4.13.0.92",
    "websocket-client>=1.8.0",
]
